Distinguish the Leaf from the Node
Now the Leat is a subclass of Node, and we can distinguish
the Leaf from the Branch.

Base on this, we make Network a subclass of Leaf which logically
make perfect sense, and then from inside an Acl class's instance,
we can access the leaves method to collect all networks of the Acl.