Implement the ACL conflict resolving method
Maintaining an orderable (can be ordered) ACL database,
is a must for the View to work properly as expected.

The following four ACLs, good1 and good2 can be ordered,
but bad1 and bad2 can not. To solve the problem, we
can split bad1 or bad2 into two separate smaller ACLs.
acl "good1" {
    192.168.1.0/24;
    10.1.1.0/24;
};
acl "good2" {
    192.168.0.0/16;
    10.1.0.0/16;
};
acl "bad1" {
    192.168.0.0/16;
    10.1.1.0/24;
};
acl "bad2" {
    192.168.1.0/24;
    10.1.0.0/16;
};

To correct and reformat an existing ACL database, these
are the rules we apply here:
    1. Every network name shall be unique
    2. A network can overlap other networks
    3. Every ACL name shall be unique
    4. ACL can be nested
    5. ACL A can cover parts of ACL B, provided that
       no part of A covered by B. If this is
       impossible, split A or B.
    6. When split an ACL, move some networks out of the
       original ACL, and put them into a new created
       one. The bottom line is move the networks which
       cause least efforts and result in least ACLs.

Changes introduced in this commit:
- Update AclGroup.addAcl method (algorithm core)
- Add AclGroup.splitAclTree method (algorithm core)
- Update class AclGroup's comment
- Update AclGroup.load method
- Add method AclGroup.parentsOfNets
- Add AclGroup.removeConflicts
- Update AclGroup.aclValidator method
- Update AclGroup.coexist method
- Add Node.rename method
- Add Node.parents method