Add method ViewGroup.insertView
Find a good location in the self.outData, and
insert the view into it.

This is the core mechanism that ensures a logically
correct view database. The rule is: IF ANY VIEW'S
ACL OVERLAPS OTHER VIEW'S, THE VIEW OF THE LESS ACL
SHALL BE PLACED FIRST, THEN THE GREATER ONE.

If it's impossible to pick a location that complies
to the order rule, raise an exception. It's possible
that the exception raised halfway at which point
some views may had already been moved from their
original position, thus corrupt the view group, to
prevent it, we make shallow copies of the groups,
and process the copies, after all existing views had
been processed, we update the view group with the
processed copies.

If a list in the orderedGroups has a view LESS or
GREATER than the newView, the same will be deleted,
all views in it will be moved to a new list. If a
view in the freeViews group LESS or GREATER than
the newView, the same will be moved to a new list.